Cleaned up sign up components
