+++ src/components/signUpForm/styles/formStyles.js
